V-790.V Phase 3.A: WTFLogAlways format-string type fix (wave 29-125)
Phase 3 initial build (PID 97339) failed at -Werror,-Wformat:

  FontCascadeCoreText.cpp:645:42: error: format specifies type
  'double' but the argument has type 'uint16_t' (aka 'unsigned
  short') [-Werror,-Wformat]
    "pt=%.1f pos=%u",

The format string used %.1f (printf for double) but ptSize at
this call site is uint16_t (see line 528).

Fix: change format to %u for the integer-typed args and cast
explicitly to unsigned/int to silence -Wformat at the argument
level too:

  - pt=%.1f → pt=%u
  - explicit casts: static_cast<int>(ane_routed),
    static_cast<unsigned>(fontId/ptSize/positionClass)

Zero behavior change — only affects log line formatting.

Re-fired build as PID 1644.

+++ Source/WebCore/platform/graphics/coretext/FontCascadeCoreText.cpp
@@ -639,10 +639,12 @@ void FontCascade::drawGlyphs(GraphicsContext& context, const Font& font, std::sp
                     // draws via CGContextDrawImage.
                     WTFLogAlways("[V-790.V] LayerB predicted "
                                  "inference_ms=%.3f ane=%d for font_id=%u "
-                                 "pt=%.1f pos=%u",
+                                 "pt=%u pos=%u",
                                  prediction->inference_ms,
-                                 prediction->ane_routed,
-                                 fontId, ptSize, positionClass);
+                                 static_cast<int>(prediction->ane_routed),
+                                 static_cast<unsigned>(fontId),
+                                 static_cast<unsigned>(ptSize),
+                                 static_cast<unsigned>(positionClass));
                 }
             }
         }
